Improve optimizeCatchBlocks()
Summary: Storing a KindOfNull is cheaper than storing other constants which may have both a type and a value that need to be written. Additionally by reordering things in fullDCE we may be able to eliminate some values whose only uses are DecRefs and StStks inside of catch traces by recording StStk uses in `optimizeCatchBlocks()`.

Reviewed By: oulgen

Differential Revision: D28290731

fbshipit-source-id: 41582fa264fd80f6d8ac088d17f4fa0730f7337d