Avoid dereferencing layout on allocation
Summary:
We shouldn't pass a bespoke::Layout* to these allocation helpers. Instead, we precompute each StructDict's m_extra field (which is a function of layout alone), and provide (sizeIndex, extra) to the helpers directly.

We could even drop extraInitializer as a param and set it in the JIT. That would save a handful of instructions (note: no loads or stores), but I'm squeamish about a native helper result that doesn't pass checkInvariants().

Reviewed By: colavitam

Differential Revision: D28366813

fbshipit-source-id: e7ce5a832e4b3d20366f1caaf543c333d48870c0