Flag to raise warning when accessing abstract context constant that does not have a default
Summary:
Currently there is a sampled warning when an abstract context constant (without default) is directly accessed in a way that would be a Hack error. However, the `m_kind` wasn't properly set on context constants, so this warning doesn't fire.

This flag adds an option to turn on the warning unconditionally instead of sampling.

Reviewed By: oulgen

Differential Revision: D28450007

fbshipit-source-id: 7508edde301282295104bc8e01a94c7949aa012d