IR for StructDict perfect hashing
Summary:
Implement perfect hash tables for StructDicts using coloring information. The
procedure for accesses in a StructDict is now as follows:
1) Check the color of the string. This will be junk for non-static strings.
2) Load the corresponding entry in the StructDict's hash table.
3) Check if the string in the entry matches the key we're querying. If so, use
the type and value index in the entry to index into the StructDict and obtain
the value.
4) Otherwise, check if the key is static. If it is, return Uninit. Otherwise,
fallback to the non-static query routine.

This is implemented in StructDict::NvGetStr and a new IR operation,
StructDictGetWithColor.

Differential Revision: D28432159

fbshipit-source-id: c4b8235f41781b6d3c994d4acfd460e5374a4fc6