Fix bug in vasm-graph-color colorer logic
Summary:
If an instruction is def-ing a PhysReg, but that PhysReg is currently
assigned to a Vreg, that Vreg must be "evicted" into another
PhysReg. The spiller has already guaranteed that it is possible to do
so (it will have spilled to maintain a possible free PhysReg). When we
evict the Vreg, we insert a copy before the current instruction,
moving it from the old PhysReg to the new one.

However, this moving logic was not taking into account any uses of the
instruction. Vreg uses are fine, but the instruction might use a
PhysReg. It is not valid to move the Vreg to a PhysReg that is used by
the instruction. That's because the inserted copy will clobber the
intended value of that PhysReg.

Remove PhysReg uses of the current instruction from the candidate set
of destination PhysRegs when moving.

Reviewed By: mofarrell

Differential Revision: D28596796

fbshipit-source-id: ac50fae7cc009bb90b48769259763c5164b7dbb9