add `Hole` for string append
Summary: This diff adds `Hole`s for ill-typed use of the string append operator, `.`. In the case the global option `TypecheckerOptions.enable_strict_string_concat_interp` is true, this means making a direct call to `Typing_ops.sub_type_res`; if false, we make use of a new version of `Typing_substring.sub_string` which returns the types at which subtyping failed, if in fact it did.

Reviewed By: madgen

Differential Revision: D28351842

fbshipit-source-id: 619b4d72e29916451ad3092b326df17b84bc3f1d