Ban static method calls on reified generics
Summary:
Although it works in the runtime, we probably want to ban static method calls on reified generics in Expression Trees. However, we only want to ban these when used within the Expression Tree expressions. We do not want to ban these when they are used in spliced expressions.

So the following is banned

```
class Foo {
  public static async function bar(ExampleContext $_):
    Awaitable<ExprTree<Code, Code::TAst, (function(): void)>>
  {
    throw new Exception();
  }
}

function takes_reified<reify T as Foo>(): void {
  Code`T::bar()`;
}
```

The following is not

```
class Foo {
  public static function bar(): ExprTree<Code, Code::TAst, (function(): void)> {
    throw new Exception();
  }
}

function takes_reified<reify T as Foo>(): void {
  Code`${ T::bar() }`;
}

```

Reviewed By: Wilfred

Differential Revision: D28374855

fbshipit-source-id: 39e3f87298fa9fcaa79e32bf128ab0b306f7ce8b