Set in_expr_tree flag only for virtualized expr
Summary:
Previously, we set the `in_expr_tree` flag for all of expression trees, meaning that we typed using this flag in both the splices (where they are extracted) and also in the desugared code, as well as the virtualized expression. However, because we are using this error primarily to give better errors based on the virtualized expression, this leads to some errors that are a bit incorrect. For instance, in the following bit of Hack code:

```
class Foo {
  public function boo(): void {}
}

function test(): void {
  $x = new Foo();
  Code`${ $x->bar() }`;
}
```

`bar` is not a valid method on `Foo`. We should report an error like:
```
No instance method `bar` in `Foo` (Typing[4053])
  Did you mean `boo` instead?
```

However, because we set the `in_expr_tree` flag on all of the expression tree, we report the following error instead:
```
`Foo` does not support this operator (it has no instance method named `bar`) (Typing[4053])
```

While this is not obviously right, this is probably not the behavior we want in Expression Trees. The fix is to only set the `in_expr_tree` flag to true when we are typing the virtualized portion of Expression Trees.

Note:
Since we might want to throw special errors based on whether or not the Visitor supports a certain Hack construct (like if a Visitor didn't want to support while loops), we may want to have separate flags for both the desugared portion and the virtualized portion?

Reviewed By: Wilfred

Differential Revision: D28362818

fbshipit-source-id: 4c166cff9ddaa54d86ea4b3265e8e2704c1dc5a5