Update closure return type
Summary:
We were using the declared type as the result type of a lambda despite inferring the correct return type and embedding it in the AST node. This change overrides the declared types return type so that the TAST and `expr` results match the return type inferred for the lambda.

From what I can tell this wasn't a typehole (perhaps because we always pull the return type of lambda out of the AST rather than the result of the recursive call to `expr`). However, this was causing us placing the wrong type (in particular `Tany`s) in the TAST and subsequently on hover types.

Reviewed By: andrewjkennedy

Differential Revision: D27794804

fbshipit-source-id: 6214d251e2417eef35d3f0b7bab01e46f61156b4