suppress scuba for all buck invocations
Summary:
Hack has the behavior that if environment variable HH_TEST_MODE is set, then it suppresses HackEventLogger. We do this because we don't want tests to generate telemetry signals. We manually write our tests to set that environment variable.

This diff takes it a step further. It says that if anyone used https://buck.build/ as their tool to invoke hack, then that was likely in a test or CI role, and should again suppress HackEventLogger.

Reviewed By: madgen

Differential Revision: D28712405

fbshipit-source-id: 4128371b500dc837d5b509451b64a2588bc01f20