Add an error to keep track of invariant violations
Summary:
When a typechecker invariant is violated, we quietly recover from it often not performing the intended check. This is neither reported to the user nor it is logged.

I created a new error for such cases. It is much like typechecker exceptions except that it doesn't log exceptions, but arbitrary telemetry.

Reviewed By: ljw1004

Differential Revision: D28676709

fbshipit-source-id: e28598019f56325c7ff646dfa9dc59edfd9823e9