Enable non-stacked inlining cost limit and retune a couple of thresholds
Summary:
This diff replaces the inlining cost limit based on the total cost of the
inlined call stack with a non-stacked cost limit (i.e., the limit to inline a
callee is just based on the callee's size).  This diff also reduces the maximum
call depth allowed to be inlined and bumps the maximum bytecode instructions in
a region.

Reviewed By: mofarrell

Differential Revision: D28738412

fbshipit-source-id: 01e01f79b101e96d3e7bda2b3014286e2eebf457