Add a type for the negation of primitive types
Summary:
When a type test returns false, refine the environment with a type that represents the negation of the type being tested for. This supports the following idiom
```
function f(num $n) : void {
  if ($n is int) {
    // $n has type int
  } else {
    // $n has type float (previously num)
  }
  // $n has type num
}
```

In this case, we could work around the old limitation by adding an additional `if ($n is float)`. That is not possible in general, since there are types that cannot be tested for.
```
function f(int | vec<int> $v) : void {
  if ($v is int) {
    // $v has type int
  } else {
    // $v has type vec<int> (previously int | vec<int>)
  }
  // $v has type int | vec<int>
}
```

This gets the basics working where primitive types can be negated (internally: they are not user denotable). In the future, support for class-ish types should be added. It would also be good to refactor and remove the primitive nonnull type from the internals, since it is equivalent to the negation of `null`.

Negations also improve the ability of the intersection code to solve types to `nothing`. `t1 <: not t2` implies that `t1 & t2 = nothing`, and giving more types precise negations, this will happen more often.

Lastly, some error reasons can be slightly degraded. Consider the first example above: after the if statement `$n : num` by unioning together the type of `$n` in both branches. Previously, the `num` in the `else` branch was the one with the reason from the type hint, and that was preserved by the union. Now, the type hint's `num` is lost in the refinement, and a new `num` is created by the union with a different reason. The same thing can happen with `mixed` and `arraykey`.

Reviewed By: CatherineGasnier

Differential Revision: D22698187

fbshipit-source-id: 9e6175dc89d13900a7ac9212f259194c52fd3b1a