Support method attributes in the DB
Summary:
This adds a new table to the native Facts DB in order to store and retrieve method attributes.

## Background

Hack allows us to decorate symbols with attributes. We use these attributes with Facts in order to organize our codebase.

We decorate many of our classes, functions, and methods with a `<<GraphQL('graphql_name')>>` attribute. We rely on Facts to index our code by these attributes, so that our build system can then generate the plumbing needed to make GraphQL work.

```
<<GraphQL('friends')>>
final class FriendsConnection implements IGraphQLConnection {}
```

Facts supports attributes on top-level types and functions, but devs are clamoring for Facts to support attributes on the methods inside of classes as well.

Reviewed By: DavidSnider

Differential Revision: D28559640

fbshipit-source-id: a738c39364030062abf8d28171ccde6fb1707464