Templatize TypeAttributeMap into `AttributeMap<Type>` to make way for `AttributeMap<Method>`
Summary:
`TypeAttributeMap` does most of the stuff we need - we just need to make it a bit more generic so we can use it to store method attributes as well as type attributes.

## Background

Hack allows us to decorate symbols with attributes. We use these attributes with Facts in order to organize our codebase.

We decorate many of our classes, functions, and methods with a `<<GraphQL('graphql_name')>>` attribute. We rely on Facts to index our code by these attributes, so that our build system can then generate the plumbing needed to make GraphQL work.

```
<<GraphQL('friends')>>
final class FriendsConnection implements IGraphQLConnection {}
```

Facts supports attributes on top-level types and functions, but devs are clamoring for Facts to support attributes on the methods inside of classes as well.

Reviewed By: DavidSnider

Differential Revision: D28559638

fbshipit-source-id: 8c54672735152330da7c9c932d0bc9079467a6f7