Add HH\Facts\method_attributes() and HH\Facts\methods_with_attribute()
Summary:
This adds builtin methods to HHVM allowing you to query the method attributes in your codebase.

## Background

Hack allows us to decorate symbols with attributes. We use these attributes with Facts in order to organize our codebase.

We decorate many of our classes, functions, and methods with a `<<GraphQL('graphql_name')>>` attribute. We rely on Facts to index our code by these attributes, so that our build system can then generate the plumbing needed to make GraphQL work.

```
<<GraphQL('friends')>>
final class FriendsConnection implements IGraphQLConnection {}
```

Facts supports attributes on top-level types and functions, but devs are clamoring for Facts to support attributes on the methods inside of classes as well.

Reviewed By: DavidSnider

Differential Revision: D28559639

fbshipit-source-id: b5ae0f5febfd0b291baafb2d95b67695293ab07d