Allow zend_strtod usage from RTA thread
Summary:
We are encountering a crash inside collectKeyOrder, which may dispatch to
zend_strtod as a result of string comparison. When zend_strtod needs to operate
on BigInts, it uses a thread-local freelist of blocks from which these integers
are allocated. The RTA thread did not initialize this freelist, leading to
crashes.

Reviewed By: arnabde03

Differential Revision: D28784101

fbshipit-source-id: 1d854e3b64348b4e35f53ed81faf9a59f5226244