Sound handling of rigid type variable escapes
Summary:
This diff proposes new logic to deal with a long-standing source of unsoundness
in Hack. Namely, it prevents the escaping of what I called "rigid type variables".

Rigid type variables include:
1. Type parameters created for `is` and `as` checks
2. Expression-dependent types
3. In the future, maybe, runtime tags

They all have in common that they are a kind of type variable that will only unify
with itself (hence "rigid" in their name). These type variables are created
dynamically during type-checking, and propagate freely into the inference
environment. However, after typing some blocks like, the body of lambdas or
loops, the rigid type variables created in the aforementioned block need to be
cleared from the environment. If not, the type system is unsound: type holes
related to this problem can be found in the tasks linked to this diff.

This diff is a first pass and only tackles lambdas, loops are less critical because
it is less easy to write type holes for them, due to the way we currently iterate
over their body.

It is sometimes safe to not eliminate rigid type variables when typing a lambda,
for example, if the lambda is called right after being created. This diff currently
does not implement this refinement.

Another refinement we can implement is the following: currently we raise an
error when we need to eliminate a type in invariant position such as `X<T#1>`,
however, if `X<T#1> <: Y` and appears itself in covariant position we could
replace `X<T#1>` altogether with `Y`.

In `typing_escape.ml` I need to find if a type parameter or an expression
dependent type is "fresh from a block". The way in which I did this is borderline
shameful; however I could not find a more convincing way. In the future we
should change internal data structures so that obtaining this information is
simpler.

Example type hole fixed by this diff:
```

class Box<T> {
  public function __construct(public T $data) {}
};

<<__EntryPoint>>
function f(): string {
  $v = Vector{};
  $f = (mixed $x) ==> {
    if ($x is Box<_>) {
      $v[] = $x->data;
      $x->data = $v[0];
    }
  };
  $f(new Box(42));
  $sbox = new Box("oops");
  $f($sbox);
  return $sbox->data; // fatal, got int expected string
}
```

Reviewed By: vsiles

Differential Revision: D28477487

fbshipit-source-id: a88956563dbc28afef3218433705ae73c8960b0a