hack: update to Bytes 1.x
Summary:
This updates hh_decl and ocamlrep to use Bytes 1.x instead of Bytes 0.5. The
main difference between the two here is that Bytes 1.x no longer provides
access to a `&mut [MaybeUninit<u8>]`, and instead provides `UninitSlice`, which
is a wrapper around `MaybeUninit` that is a bit more restrictive.

The code as it exists today is "technically" violating the safety expectations
of the slices in question, since the docs say you shouldn't read from them but
we do. However, those reads only happen after initializing, so while they're
perhaps not totally in line with the letter of the safety expectations on
`UninitSlice`, they seem to follow them in spirit (I documented this inline).

Differential Revision: D28601254

fbshipit-source-id: 9a3d892947bc0ac71ba28f23e57bd24c9334713d