Fix bug: ignore SrcKeys in eval'd functions
Summary: Jumpstart doesn't handle SrcKeys in eval'd functions. It serializes units by filename, but such functions have an empty filename. That's not a problem for profiling tracelets, because we've made these units interp-only, but it's a problem for CPGO because we still serialize SrcKeys for sources and sinks that get interpreted.

Reviewed By: colavitam

Differential Revision: D28853665

fbshipit-source-id: f71238f052e8d532d479878139aba73cc8427d2b