Fix tc-print to work without a RepoFile
Summary:
A bytecode repo is not always needed/produced by hhvm anymore, so make tc-print
work even without it (it'll just not print bytecodes).  This diff also drops the
"magic" that was used to lookup bytecode repo using environment variables and
config settings, and instead just add a command-line option to tc-print to
specify the location of the repo file.

Reviewed By: binliu19

Differential Revision: D28649364

fbshipit-source-id: cd5044614941b8d225a52be3c6b27d8cb4c6f5c7