Interface Generator.
Summary:
A simple interface generator, emit Hack interface definition.

Each `_HackInterfaceGenerator` corresponding to an interface, therefore it has a `name` to track the interface name, and an `extend` to track the extend relationship with other interfaces .

Currently, we only have one `add_extend` public method to add `extend` relationship to this object.

Reviewed By: jamesjwu

Differential Revision: D28817221

fbshipit-source-id: 10f57edc12c9a8dbb0fac83c440b49275b6a70ed