Class Generator.
Summary:
A simple class generator, emit Hack class definition.

Each `_HackClassGenerator` corresponding to a class, therefore it has a `name` to track the class name, an `implements` to track the implement relationship with other interfaces, and one `extend` to track the extend relationship with parent class.

Currently, we only have `set_extend` public method to set `extend` relationship to this object, and `add_implement` to set `implement` relationship to this object.

Reviewed By: jamesjwu

Differential Revision: D28817220

fbshipit-source-id: bc6c86964e7e3f59f0d6fd3ce9f574bbd504c757