Only print the const-val string for constants
Summary:
For constant values, we should only include the const-val string in the type description, not the fully-qualified type. This change only affects the JSON tcprint output, which now matches the console output.

Minor subtlety: we check tmp->inst()->is(DefConst) instead of tmp->hasConstVal() to cover Bottom, Uninit, InitNull, and Nullptr.

Reviewed By: colavitam

Differential Revision: D28926159

fbshipit-source-id: d37f1d6158905ad7de7b46e9f1567badced24925