Emit service request stubs after relocation
Summary:
Emit service request stubs after relocation. At this point the code is at its
final location, so the logic is simpler. Error handling is improved as well,
as we are now able to handle a situation when we run out of space while
emitting the stub. It also allows us to bind into the target SrcRec, which is
deferred to a later diff.

Reviewed By: mofarrell

Differential Revision: D28921215

fbshipit-source-id: 15d9d4d3255950300b82675a57d129c9d19af3b0