Fix bug on empty, counted arrays
Summary:
If we put an empty, counted array into APC, the bespoke integration will core on attempting to make this array uncounted. We should simply use a static empty array for this case instead.

This bug is asymptomatic in prod, because MixedArray::MakeUncounted works fine on empty inputs. The assertx(!empty()) there is just a guard against perf bugs.

This diff may have a perf effect, because it eliminates extra copies of empty arrays.

Reviewed By: colavitam

Differential Revision: D28945080

fbshipit-source-id: 913387f97e9c500db22620e874a99740e8f50059