report "writing crash logs" to progress
Summary:
As discussed in the previous diff, there's 5-10 seconds delay after a TYPECHECKER_EXIT during which the monitor writes crash logs. (I'd actually like to increase this to 20s because we're failing to get crash logs a lot of the time due to timeouts, but that's a matter for a future diff). During this time, the monitor exists but isn't accepting any new connections. (We also used to be getting SERVER_PROGRESS_EXEPTION during that interval, which was addressed by the previous diff).

Option1: we could make it so any existing hh_client quits as soon as TYPECHECKER_EXIT happens, and then any future attempts to invoke "hh" over the next ten seconds will fail to make progress for no clear reason to the end user since the monitor is busy, and then subsequent attempts will succeed

Option2: we could make it so any existing hh_client displays "writing crash logs" while the monitor is doing its work. This will give transparency to the user. And once it's done, then the monitor will be ready to handle further connections.

In this diff I've opted for Option2. I've also refactored the code a little, inlining the TYPECHECKER_EXIT function.

Differential Revision: D28924555

fbshipit-source-id: a68c31a91c8386013344f1cffe36f0dacdad46c2