allow types in hsl .hhi files to be considered enforced
Summary:
Types in .hhi files are not considered enforced, because we aren't always sure if hhvm is doing the enforcement check specified in the .hhi file. However, the .hhi files in hsl and hsl_generated should not have any of these problems, and if they do, that is a bug.

This means that `dynamic` typed arguments can be used to call hsl functions that have otherwise enforceable types (`string`, `int`, `C`, etc.).

Reviewed By: vassilmladenov

Differential Revision: D28938651

fbshipit-source-id: b6b65cb2fd60307d5f8700d19ff088792e0d7724