Update null_coalesce_assignment documentation
Summary: Investigating null coalesce assignment for an hhvm optimization. Spent longer trying to determine how `null_coalesce_assignment` was implemented due to no history after we migrated emitters to new `InstrSeq`, and clever way the read-then-write behavior of it is implemented. (We leave the props/elem instructions for the member op chain on the stack in case we do a write later. Documentation would've been helpful here.

Reviewed By: shiqicao

Differential Revision: D28926351

fbshipit-source-id: 854494e8ead5ec8654bb080952c3b9a19a2d4f66