hphp buck: replace open call with generated .bzl file
Summary:
Buck Starlark does not provide `open` function to read `php.txt` in Buck files.

This diff instead generates `php.bzl` file from `php.txt`.

Buck test added to guarantee that `php.bzl` is in sync with `php.txt`.

This diff is added to be able to convert hphp to Starlark (following diff D28953252), which is is part of an effort to switch fbcode Buck setup from Python DSL to Starlark.

Reviewed By: DrMarcII

Differential Revision: D28953253

fbshipit-source-id: 3ab151c38b006af3f432d27314cfe264cd2a9700