Do not use spin lock to protect histograms, time series and inline cost cache
Summary:
Spin locks are bad in user space and should never be used. If a machine is out
of CPU and a thread holding this lock gets preempted, all other threads will be
spinning endlessly and the kernel will be giving them time slots as it can't
see they are not making progress. But since there is more demand for CPU than
available, these time slots will be shrinking and more likely to preempt right
after finally obtaining a lock.

Use a default folly::SharesMutex instead, that tries to spin for a bit before
grabbing a good old mutex.

Reviewed By: ricklavoie

Differential Revision: D29010086

fbshipit-source-id: e7aeb05e4966cc73aab4d532a997db710cde8130