Fix spilling calculation involving uses in vasm-graph-color
Summary:
When calculating spilling candidates, vasm-graph-color was
intentionally excluding the uses of the the instruction. The logic was
that since the instruction needs to use the Vreg, it has to be in a
physical register, and thus should not be a candidate for spilling.

However, it's more subtle than that. For a given instruction, we do
two spilling calculations. The first is any necessary spilling after
reloading any spilled uses. In this situation, indeed we cannot spill
any uses (for the above mentioned reason). The second spilling
calculation is for any spills triggered by writing the defs. In this
situation, we *can* spill the uses. The key is that if we do so, we
still need to keep them in physical registers until they're read, but
become dead immediately after (before the defs are written). This can
be accomplished by ssaalias to control how the SSA rewriting is done.

This is actually required in some situations. Consider:

copyargs v1 ...... -> r1 .....

Assume that r1... consists of all physical registers, and
that v1.... are all alive after the copyargs. This is not actually
colorable unless we spill the uses! The dsts of the copyargs forces
all Vregs to be spilled.

Reviewed By: mofarrell

Differential Revision: D29013903

fbshipit-source-id: 54fabf019be5c09d1207db4755a27ad901c21834