Simplify logic for private access
Summary: The logic for checking access to a private member is overly complex. The only case when the class of the member and the class of the use-site should be *different* is when there is a class-level constraint of the form `where this = C`. (And a future improvement might be to pre-instantiate `this` to `C` so that we don't need the special casing in visibility checks.)

Reviewed By: vsiles

Differential Revision: D29063967

fbshipit-source-id: e9145464008f25ad88d0db55588c773982b416ee