JIT more cases of is_a
Summary:
We can JIT all four common cases of is_a: a {Class, Object} on the LHS being tested against a {Class, String} on the RHS. All we have to do is "lift" both sides to a Class, then compare the classes. We can further optimize the InstanceOf comparison in cases where we have some info about the RHS (for example, if it's an interface).

One strange thing about the implementation here is that we use LookupClsRDS to get from a String to a Class rather than LdCls. I was initially confused about why the native helper does that - lookup, unlike load, does not invoke the autoloader - until I realized that in all cases where the result would be "true", we must have already autoloaded all superclasses and interfaces implemented by the LHS.

Reviewed By: mofarrell

Differential Revision: D29091339

fbshipit-source-id: 703379b5446ac57f5f657fbb3c55b25a0e89a6eb