Null Coalesce Assignment Exception Tests
Summary:
We'd like to make `NullCoalesce` another SetOpOp when the rhs is side-effect free. This results in different exceptions thrown due to moving from 2 member op chains (MOpMode::Warn + MOpMode::Define) to just a member op chain in MOpMode::Define.

There are no tests stressing the exceptions thrown by null coalesce assignment operator. Adding here to ensure future diffs do not change behavior and to note the current behavior.

In particular, one thing that may need to change for the optimization to work, is for const properties to throw, when they are used on the lhs of a `??=`. Currently, when using a class const on the lhs, HackC throws a fatal (even for `??=` when it may _not_ write):
```
Fatal error: Array-like class consts are not valid lvalues
```

But const props have different behavior for `??=` and `=`. This will need to change for the optimization to work, as we do not know at the prop dim `d` whether we write or not, and therefore can't throw "Mutate const prop". We then lose this information when we move to the elem op.

```
class A {
  <<__Const>> public dict<int, int> $d = dict[1 => 10];
}

$a = new A();
$a->d[1] ??= 10; //ok
$a->d[2] = 10; // fatal: Mutate const prop.

```

Reviewed By: oulgen

Differential Revision: D28983860

fbshipit-source-id: e700d2c427aeb642b6f4f87cade963ea7950a806