Introduce StringishObject, Have HHVM magically generate StringishObject instead of Stringish on objects
Summary:
As part of a way to give a roadmap for folks who are using deprecated replacements for old-style stringish_cast to replace that with reasonable code.

The idea being that they have a stringish object, they can test against this type and then safely invoke the method vs Stringish::__toString which might actually try to invoke a method on a string.

Reviewed By: vassilmladenov

Differential Revision: D28972918

fbshipit-source-id: e1fbe9e10bb346bba92c9d8eb76f1b2bdbe37d44