Remove Traversable::C, Container::CMut
Summary:
Because `Traversable` actually includes arrays which are not objects, resolving polymorphic coeffects for ctx constants off `Traversable` will be costly at runtime.

In light of this and non-Container Traversables are rarely used in WWW, we no longer plan to provide typechecking support for coeffect requirements implicit to `foreach`. These will only be enforced by the runtime.

Discussion: https://fb.workplace.com/groups/681804496083170/permalink/826543798275905

This diff only changes the hhi; it appears that we never even added them to the HHVM stubs.

Similarly for Container::CMut

Reviewed By: rodmk

Differential Revision: D28904507

fbshipit-source-id: 719211c1026ddfaf3da668511e95e7301d2806c1