Simplify logic for protected access
Summary: In order to check `protected` access for code in classish type `C` accessing a protected member in classish type `D`, we need to check that `C` is a descendant of `D` through some sequence of `extends`, `implements`, `use`, `require extends` and `require implements` steps. This doesn't actually correspond to any decl query that we implement at present, as `get_ancestor` doesn't close with respect to `requires`. So the current implementation calls out to subtyping, with some fake `Tany` generic parameters. This can be expensive. For now, implement a `has_ancestor_including_req` function that does the reflexive transitive closure of the steps listed above. We might in future want to extend the decl API with this, perhaps even using it from subtyping.

Reviewed By: vsiles

Differential Revision: D29103103

fbshipit-source-id: 4d6b00f782443c97ab985db1f42202af9d702fd0