Fix function case mismatches in hhvm tests
Summary:
One of the last remaining hurdles to making functions and classes fully
case sensitive is the hhvm test suite. This diff is a first stab at fixing low
hanging fruit, where function case was wrong in tests for other things.

If a test was obviously intentionally testing function case-insensitivity,
I left it alone.  Most of this diff is `s/hh/HH/` but there are a few other
miscellaneous changes as well.

Reviewed By: paulbiss

Differential Revision: D29104741

fbshipit-source-id: 40eb8dd74977f617b3ae5fb2a3b9653c16ab6c33