Introduce a codegen context and escape hatch codegen_unsafe
Summary:
The primary motivation is so that HHBC/HHVM can map it into defaults and
not enforce it, unlike the `write_props` context.  The context does _not_ have
additional capabilities so that it can be easily callable from the `defaults` context.

Also, introduce the appropriate escape hatches via a new unsafe context `codegen_unsafe`,
which is a more principled way to allow impurities (such as IO) not in the body but
in direct or indirect callees (see test case `.../local/codegen_unsafe.php`).
For now, allow calls to unannotated/default context.

In the future we may make the restrictions tighter by including `IO` capability and
`ReadGlobals` (especially if we ever introduce a dedicated context for the former)

Reviewed By: vassilmladenov

Differential Revision: D29003351

fbshipit-source-id: 8f6954ec53b7565968c8b9e1e9f6abb77f6e87b2