Add ability to return to JIT from the middle of interpreted FCall
Summary:
Pass a flag to FCall opcodes indicating whether the returned TCA will be used
unconditionally to reenter JIT, such as when the interpreter was invoked from
handleResume() or via interpOneCF(). Add TCA return value to FCall opcodes, but
return nullptr for now.

This will be used to continue execution in JIT at the FuncEntry position rather
than the first opcode, so that we avoid needing a separate translation for
functions entered via interpreted FCall.

Reviewed By: ricklavoie

Differential Revision: D29122793

fbshipit-source-id: 23b0e4e2c1ac412cc327906e501e13fd6859bfd5