Make assignment refinement an intersection
Summary:
We refine on assignments to properties. For example,
```
class C { public ?int $i; }
...
$i = 0;
$c->i = $i
```
creates a fake variable  `$c->i` with type `int` instead of `?int`.

This works nicely when there aren't any errors. Particularly when, the property hint is a subtype of the RHS of the assignment because `ty & ty' = ty` when `ty <: ty'`.

When there a FIXMEs, however, it doesn't work as nicely. Consider the following:
```
class C { public string $i; }
...
$i = 0;
// HH_FIXME[4110]
$c->i = $i
takes_string($c->i);
```
Now the type of the fake variable `$c->i` is `int` which leads to an error in the call to `takes_string`. So the first FIXME leads to a second error.

To avoid this, this diff changes the refinements by assignment to generate intersections instead. On the happy path, this makes no difference because `ty & ty'` with `ty <: ty'` gets simplified to `ty`, but on the error path we avoid the second error because the type is `int & string`.

We do the same  for static properties as well.

Reviewed By: mpu

Differential Revision: D28896541

fbshipit-source-id: ce1cae8e195620b7603c1cce116e47f48ea54ea4