Send failed actions back into action queue
Summary:
Adds capability to partition files into buckets to send to Remote Execution for type-check, each in the form of the following action
```
hh --config use_mini_state=true \
    --config prechecked_files=true \
    --config load_state_natively_v4=true \
    --allow-non-opt-build \
    --ignore-hh-version
    --multi-remote-execution <bucket>
```
Currently ~15-25% of actions fail with "Hh_server is overloaded with too many concurrent requests" on the remote side. When hulk fails to parse such an action result, it crashes. This diff makes hulk rerun these failed actions instead of crashing on the local side. Based on what I've seen, the failed actions all succeed with the first retry.

Note: In the long run we shouldn't throw away the whole session if a few actions fail, hence this change. For now it serves as a way to get verifiable hulk runs while we investigate the 15-25% failures.

Reviewed By: bobrenjc93

Differential Revision: D29073171

fbshipit-source-id: d3fd50b53c1d00b7eec3d5018ebdd5c9687f45e8