Add unit file cache extension
Summary:
Add a new extension called "unit file cache". This extension uses the
existing unit emitter cache machinery to store cached UnitEmitters in
a SQLlite file. This extension serves two purposes. First, it is a
replacement for the old Sandbox HHBC repo (which is about to be
removed). Second, it serves as an example of how to use the cache hook
mechanism to provide your own caching.

The extension is disabled by default. To enable it, simply specify
UnitFileCache.Path to the file path (placeholders accepted). The
extension is "best effort", it simply logs errors and keeps going. The
extension can also be used during RepoAuthoritative builds to speed up
parsing.

For test coverage, enable it for all tests run by hphp/test/run.php,
including the repo tests. As a nice side-effect, this speeds up test
runs by quite a bit (by avoiding redundant parsing).

Reviewed By: jano

Differential Revision: D29180851

fbshipit-source-id: 9b8e0dea0bdb10245dab2276e09558564d5775ad