simplify Diagnostic_subscription now that errors all fire in the proper file
Summary:
The code in Diagnostic_subscription subscription makes sure to eliminate errors whose primary error position are not in the current file, IOW whose "source" does not correspond to the file of the first line of the error.

Since this logic was implemented, a workaround to primary error position bugs has been implemented in Errors.check_pos_msg which wraps such errors so that they fire in the proper file. A lot of logic in Diagnostic_subscription is therefore not useful anymore.

The crux of this diff is essentially to replace
```
local_errors: RP.Set.t RP.Map.t;
```
with
```
diagnosed_files: RP.Set.t;
```
with semantics `diagnosed_files = RP.Map.keys local_errors`

Most changes in this diff follow from that change.

Reviewed By: dabek

Differential Revision: D28997450

fbshipit-source-id: 613adea13ded803568b5762f9765d882a183d2ed