Fix BBC bug with SetOpL, math, and exceptions
Summary:
SetOpL with Add/Sub/Mul works by first converting the LHS to num and then doing the actual op. In a case like the following

```
  $a = null;
  $e = keyset[];
  try {
    $a += $e;
  } catch (Exception $e) {
    var_dump($a);
  }
```

This results in $a being `0` in the exception block. However, we didn't propogate this correctly in hhbbc. We assumed that the op had no effects whatsoever when an exception is thrown, meaning $a could be replaced with `null` in the catch block.

I've updated BBCs logic to handle this situation correctly.

Differential Revision: D29190295

fbshipit-source-id: 20d7c29367f7ab346fb86806e78b757454abbb8f