Graph generator ASP program and Python program based on given parameter.
Summary:
Adding a new Logic Programming file, `graph_generator.lp`, which can be used for creating a dependency graph based on the parameters given by the user.

Added `generate_logic_rules` mainly creates a bunch of internal symbol representations that can be used for easy reasoning (the number) and easy code emitting (the string). Also, this function will handle the `min_depth` requirement by using normal distribution to create a **backbone** of the synthesized code base, later on the reasoning engine can expand the backbone. This idea is imitating the software development in real world. Later, we can experiment more ideas once we got more profiling result.

That's why we have predicate `internal_symbols("S0", 0).` due to the actual Hack/Java/C# code restricts the Class/Interface/Function name can't start with a number, but numeric computation during the reasoning requires a number.

We are assuming the parameters given by the user are valid, otherwise the Clingo can produce "Unsatisfiable" result, in that case, the emitted code will be empty.

The good thing about declarative programming is we can easily mixed up the implementations in any order, as long as there is no parameter given, this part of knowledge is remaining hibernate during the reasoning process.

Reviewed By: madgen

Differential Revision: D29056607

fbshipit-source-id: 733a2cd6381d19ef612b08a69954bcf60a18f1bc